refactor settings page
